requirements.txt
gensim==3.5.0
numpy==1.16.3
tensorflow-gpu==2.1
sacrebleu==1.4.3
setuptools==41.0.1
spacy==2.2.2
tqdm==4.40.2